add mileage consideration. parsed optimizedPurchase module
